pyproject.toml
[project]
name = "eolo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ultralytics[my-extra]",
    "rootutils>=1.0.7",
    "hydra-core==1.3.2",
    "hydra-colorlog==1.2.0",
    "wandb>=0.19.8"

]

[tool.uv.sources]
ultralytics = { path = "../EOLO-ultralytics", editable = true }